# 2W1-G3-VENTAS/ventas-api/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'ventas-api'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/ventas/{id}":{"get":{"tags":["venta-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VentaResponse"}}}}}},"put":{"tags":["venta-controller"],"operationId":"putEstado","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"estado","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VentaResponse"}}}}}},"delete":{"tags":["venta-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VentaResponse"}}}}}}},"/ventas/facturacion/{id}":{"get":{"tags":["venta-integration-controller"],"operationId":"getVentaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntegrationVentaResponse"}}}}}},"put":{"tags":["venta-integration-controller"],"operationId":"actualizarEstado","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"state","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/ventas/":{"get":{"tags":["venta-controller"],"operationId":"getByFiltros","parameters":[{"name":"id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"id_cliente","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"id_vendedor","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"fecha_min","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"fecha_max","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"forma_entrega","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"tipo_venta","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"monto_desde","in":"query","required":false,"schema":{"type":"number"}},{"name":"monto_hasta","in":"query","required":false,"schema":{"type":"number"}},{"name":"estado","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VentaResponse"}}}}}}},"post":{"tags":["venta-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VentaRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VentaResponse"}}}}}}},"/ventas/all/":{"get":{"tags":["venta-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VentaResponse"}}}}}}}},"/reportes":{"get":{"tags":["reporte-controller"],"operationId":"generateReport","parameters":[{"name":"mes","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"anio","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"tipo_venta","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReportResponse"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"DescuentoResponse":{"type":"object","properties":{"descripcion":{"type":"string"},"monto":{"type":"number"}}},"DetalleVentaResponse":{"type":"object","properties":{"cod_producto":{"type":"string"},"descripcion":{"type":"string"},"precio_unitario":{"type":"number"},"cantidad":{"type":"integer","format":"int32"}}},"VentaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fecha":{"type":"string","format":"date-time"},"doc_cliente":{"type":"integer","format":"int64"},"tipo_venta":{"type":"integer","format":"int32"},"forma_entrega":{"type":"integer","format":"int32"},"fecha_entrega":{"type":"string","format":"date-time"},"id_vendedor":{"type":"integer","format":"int64"},"detalles":{"type":"array","items":{"$ref":"#/components/schemas/DetalleVentaResponse"}},"descuentos":{"type":"array","items":{"$ref":"#/components/schemas/DescuentoResponse"}},"subtotal":{"type":"number"},"total":{"type":"number"},"estado":{"type":"integer","format":"int32"},"id_reserva":{"type":"integer","format":"int64"}}},"DescuentoRequest":{"required":["descripcion","monto"],"type":"object","properties":{"descripcion":{"type":"string"},"monto":{"type":"number"}}},"DetalleVentaRequest":{"required":["cantidad","cod_producto","descripcion","precio_unitario"],"type":"object","properties":{"cod_producto":{"type":"string"},"descripcion":{"type":"string"},"precio_unitario":{"type":"number"},"cantidad":{"type":"integer","format":"int32"}}},"VentaRequest":{"required":["detalles","doc_cliente","fecha","fecha_entrega","forma_entrega","id_vendedor","tipo_venta"],"type":"object","properties":{"fecha":{"type":"string","format":"date-time"},"doc_cliente":{"type":"integer","format":"int64"},"tipo_venta":{"maximum":2,"minimum":1,"type":"integer","format":"int32"},"forma_entrega":{"maximum":2,"minimum":1,"type":"integer","format":"int32"},"fecha_entrega":{"type":"string","format":"date-time"},"id_vendedor":{"type":"integer","format":"int64"},"detalles":{"type":"array","items":{"$ref":"#/components/schemas/DetalleVentaRequest"}},"descuentos":{"type":"array","items":{"$ref":"#/components/schemas/DescuentoRequest"}}}},"IntegrationDescuentoResponse":{"type":"object","properties":{"monto":{"type":"number"},"descripcion":{"type":"string"}}},"IntegrationDetalleVentaResponse":{"type":"object","properties":{"cod_producto":{"type":"string"},"descripcion":{"type":"string"},"precio_unitario":{"type":"number"},"cantidad":{"type":"integer","format":"int32"}}},"IntegrationVentaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"id_reserva":{"type":"integer","format":"int64"},"fecha":{"type":"string","format":"date-time"},"doc_cliente":{"type":"integer","format":"int64"},"detalles":{"type":"array","items":{"$ref":"#/components/schemas/IntegrationDetalleVentaResponse"}},"descuentos":{"type":"array","items":{"$ref":"#/components/schemas/IntegrationDescuentoResponse"}},"subtotal":{"type":"number"},"total":{"type":"number"}}},"ReportProductResponse":{"type":"object","properties":{"descripcion":{"type":"string"},"cantidad":{"type":"integer","format":"int64"}}},"ReportResponse":{"type":"object","properties":{"monto":{"type":"number"},"total_ventas":{"type":"integer","format":"int32"},"monto_mes_anterior":{"type":"number"},"total_mes_anterior":{"type":"integer","format":"int32"},"mayorista":{"type":"integer","format":"int32"},"minorista":{"type":"integer","format":"int32"},"pendiente":{"type":"integer","format":"int32"},"entregado":{"type":"integer","format":"int32"},"productos":{"type":"array","items":{"$ref":"#/components/schemas/ReportProductResponse"}}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}