# 2W1-G3-VENTAS/ventas-api/src/main/java/ar/edu/utn/frc/tup/lciv/models/dtos/integration/IntegrationDescuentoResponse.java
package ar.edu.utn.frc.tup.lciv.models.dtos.integration;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class IntegrationDescuentoResponse {
    private BigDecimal monto;
    private String descripcion;
}