# 2W1-G3-VENTAS/ventas-api/src/main/java/ar/edu/utn/frc/tup/lciv/models/dtos/responses/DescuentoResponse.java
package ar.edu.utn.frc.tup.lciv.models.dtos.responses;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class DescuentoResponse {
    private String descripcion;
    private BigDecimal monto;
}